Add basic fxns to try iNat servers up to 3 times
This helps overcome server timeouts and other non-fatal errors. Does not catch or overcome fatal errors at the moment. Note, however, that harvesting occurs sequentially and independently. Thus, if an error is thrown during a batch harvest, the harvest can resume with the next property. The harvesting of observations on properties prior to the error is not affected/undone.